Fix grid resize feature
Create resizeBoxes function that changes the height and width of the
boxes depending on the number of boxes so that the grid is always
square. Added that funciton to the newGrid function.

At the moment, the clear and fill functions revert the boxes
to their original size. Maybe something to do with css.